requirements.txt
aioschedule==0.5.2
aiosmtplib==1.1.7
anyio==3.6.2
bcrypt==4.0.1
beanie==1.15.4
blinker==1.5
boto3==1.26.23
botocore==1.29.23
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.88.0
fastapi-mail==1.1.5
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
motor==3.1.1
multidict==6.0.3
numpy==1.23.5
oauth2client==4.1.3
openpyxl==3.0.10
pandas==1.5.2
passlib==1.7.4
Pillow==9.3.0
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toml==0.10.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
XlsxWriter==3.0.3
yarl==1.8.2
